[Show] Rename to uppercase Components/ folder
