+++ src/v2/Apps/Show/Components/ShowInstallShots.tsx
