+++ src/v2/Apps/Show/ShowApp.tsx
@@ -6,14 +6,14 @@ import { HorizontalPadding } from "v2/Apps/Components/HorizontalPadding"
 import { Footer } from "v2/Components/Footer"
 import { ErrorPage } from "v2/Components/ErrorPage"
 import { Box, Column, GridColumns, Separator } from "@artsy/palette"
-import { ShowMetaFragmentContainer as ShowMeta } from "v2/Apps/Show/components/ShowMeta"
-import { ShowHeaderFragmentContainer as ShowHeader } from "./components/ShowHeader"
-import { ShowAboutFragmentContainer as ShowAbout } from "./components/ShowAbout"
-import { ShowInstallShotsFragmentContainer as ShowInstallShots } from "./components/ShowInstallShots"
-import { ShowContextualLinkFragmentContainer as ShowContextualLink } from "./components/ShowContextualLink"
-import { ShowViewingRoom } from "./components/ShowViewingRoom"
+import { ShowMetaFragmentContainer as ShowMeta } from "v2/Apps/Show/Components/ShowMeta"
+import { ShowHeaderFragmentContainer as ShowHeader } from "./Components/ShowHeader"
+import { ShowAboutFragmentContainer as ShowAbout } from "./Components/ShowAbout"
+import { ShowInstallShotsFragmentContainer as ShowInstallShots } from "./Components/ShowInstallShots"
+import { ShowContextualLinkFragmentContainer as ShowContextualLink } from "./Components/ShowContextualLink"
+import { ShowViewingRoom } from "./Components/ShowViewingRoom"
 import { ShowApp_show } from "v2/__generated__/ShowApp_show.graphql"
-import { ShowArtworksRefetchContainer as ShowArtworks } from "./components/ShowArtworks"
+import { ShowArtworksRefetchContainer as ShowArtworks } from "./Components/ShowArtworks"
 
 interface ShowAppProps {
   show: ShowApp_show
